Add create_test_table fixture factory; extract transaction tests
This mostly reverts test_connection.py back to starting state
