+++ tests/fix_db.py
@@ -65,3 +65,26 @@ def svcconn(dsn):
     conn = Connection.connect(dsn, autocommit=True)
     yield conn
     conn.close()
+
+
+@pytest.fixture
+def create_test_table(svcconn):
+    """
+    This fixture gives a factory for creating tables for use in tests.
+
+    NB: Should be specified as the first fixture in test method signatures in
+    order to ensure that it gets cleaned up last, after any other connections
+    used in the test are closed (so that they can't block dropping the tables.)
+    """
+    tables = []
+
+    def create(name: bytes, columns: bytes):
+        assert name not in tables, f"Table {name} already created in this test"
+        tables.append(name)
+        svcconn.pgconn.exec_(b"drop table if exists %b" % name)
+        svcconn.pgconn.exec_(b"create table %b (%b)" % (name, columns))
+
+    yield create
+
+    for name in tables:
+        svcconn.pgconn.exec_(b"drop table %b" % name)
