+++ tests/test_connection.py
@@ -1,5 +1,4 @@
 import gc
-from contextlib import contextmanager
 
 import pytest
 import logging
@@ -11,18 +10,8 @@ from psycopg3.conninfo import conninfo_to_dict
 
 
 @pytest.fixture
-def temp_table(svcconn):
-    """
-    Creates a table called 'temp_table' for use in tests.
-
-    NB: Should be specified as the first fixture in test method signatures in
-    order to ensure that it gets cleaned up last, after any other connections
-    used in the test are closed (so that they can't block dropping this table.)
-    """
-    svcconn.pgconn.exec_(b"drop table if exists temp_table")
-    svcconn.pgconn.exec_(b"create table temp_table (id text primary key)")
-    yield
-    svcconn.pgconn.exec_(b"drop table temp_table")
+def table_foo(create_test_table):
+    create_test_table(b"foo", b"id int primary key")
 
 
 def test_connect(dsn):
@@ -62,27 +51,27 @@ def test_weakref(dsn):
     assert w() is None
 
 
-def test_commit(temp_table, conn):
+def test_commit(table_foo, conn):
     conn.pgconn.exec_(b"begin")
     assert conn.pgconn.transaction_status == conn.TransactionStatus.INTRANS
-    conn.pgconn.exec_(b"insert into temp_table values ('foo')")
+    conn.pgconn.exec_(b"insert into foo values (1)")
     conn.commit()
     assert conn.pgconn.transaction_status == conn.TransactionStatus.IDLE
-    res = conn.pgconn.exec_(b"select id from temp_table where id = 'foo'")
-    assert res.get_value(0, 0) == b"foo"
+    res = conn.pgconn.exec_(b"select id from foo where id = 1")
+    assert res.get_value(0, 0) == b"1"
 
     conn.close()
     with pytest.raises(psycopg3.OperationalError):
         conn.commit()
 
 
-def test_rollback(temp_table, conn):
+def test_rollback(table_foo, conn):
     conn.pgconn.exec_(b"begin")
     assert conn.pgconn.transaction_status == conn.TransactionStatus.INTRANS
-    conn.pgconn.exec_(b"insert into temp_table values ('foo')")
+    conn.pgconn.exec_(b"insert into foo values (1)")
     conn.rollback()
     assert conn.pgconn.transaction_status == conn.TransactionStatus.IDLE
-    res = conn.pgconn.exec_(b"select id from temp_table where id = 'foo'")
+    res = conn.pgconn.exec_(b"select id from foo where id = 1")
     assert res.ntuples == 0
 
     conn.close()
@@ -90,24 +79,24 @@ def test_rollback(temp_table, conn):
         conn.rollback()
 
 
-def test_auto_transaction(temp_table, conn):
+def test_auto_transaction(table_foo, conn):
     cur = conn.cursor()
     assert conn.pgconn.transaction_status == conn.TransactionStatus.IDLE
 
-    cur.execute("insert into temp_table values ('foo')")
+    cur.execute("insert into foo values (1)")
     assert conn.pgconn.transaction_status == conn.TransactionStatus.INTRANS
 
     conn.commit()
     assert conn.pgconn.transaction_status == conn.TransactionStatus.IDLE
-    assert cur.execute("select * from temp_table").fetchone() == ("foo",)
+    assert cur.execute("select * from foo").fetchone() == (1,)
     assert conn.pgconn.transaction_status == conn.TransactionStatus.INTRANS
 
 
-def test_auto_transaction_fail(temp_table, conn):
+def test_auto_transaction_fail(table_foo, conn):
     cur = conn.cursor()
     assert conn.pgconn.transaction_status == conn.TransactionStatus.IDLE
 
-    cur.execute("insert into temp_table values ('foo')")
+    cur.execute("insert into foo values (1)")
     assert conn.pgconn.transaction_status == conn.TransactionStatus.INTRANS
 
     with pytest.raises(psycopg3.DatabaseError):
@@ -119,7 +108,7 @@ def test_auto_transaction_fail(temp_table, conn):
 
     conn.commit()
     assert conn.pgconn.transaction_status == conn.TransactionStatus.IDLE
-    assert cur.execute("select * from temp_table").fetchone() is None
+    assert cur.execute("select * from foo").fetchone() is None
     assert conn.pgconn.transaction_status == conn.TransactionStatus.INTRANS
 
 
@@ -351,291 +340,3 @@ def test_notify_handlers(conn):
 
     with pytest.raises(ValueError):
         conn.remove_notify_handler(cb1)
-
-
-def test_transaction(conn):
-    """Basic use of transaction() to BEGIN and COMMIT a transaction."""
-    assert conn.pgconn.transaction_status == conn.TransactionStatus.IDLE
-    with conn.transaction():
-        assert conn.pgconn.transaction_status == conn.TransactionStatus.INTRANS
-    assert conn.pgconn.transaction_status == conn.TransactionStatus.IDLE
-
-
-def test_transaction_exposes_associated_connection(conn):
-    """Transaction exposes its connection as a read-only property."""
-    tx = conn.transaction()
-    assert tx.connection is conn
-    with pytest.raises(AttributeError):
-        tx.connection = conn
-
-
-def test_transaction_begins_on_enter(conn):
-    """Transaction does not begin until __enter__() is called."""
-    tx = conn.transaction()
-    assert conn.pgconn.transaction_status == conn.TransactionStatus.IDLE
-    with tx:
-        assert conn.pgconn.transaction_status == conn.TransactionStatus.INTRANS
-    assert conn.pgconn.transaction_status == conn.TransactionStatus.IDLE
-
-
-def test_transaction_commit_on_successful_exit(temp_table, conn):
-    """Changes are committed on successful exit from the `with` block."""
-    cur = conn.cursor()
-    with conn.transaction():
-        cur.execute("insert into temp_table values ('foo')")
-
-    assert conn.pgconn.transaction_status == conn.TransactionStatus.IDLE
-    assert cur.execute("select * from temp_table").fetchone() == ("foo",)
-
-
-def test_transaction_rollback_on_exception_exit(temp_table, conn):
-    """Changes are rolled back if an exception escapes the `with` block."""
-    cur = conn.cursor()
-    with pytest.raises(ExpectedException):
-        with conn.transaction():
-            cur.execute("insert into temp_table values ('foo')")
-            raise ExpectedException("This discards the insert")
-
-    assert conn.pgconn.transaction_status == conn.TransactionStatus.IDLE
-    assert cur.execute("select * from temp_table").fetchone() is None
-
-
-def test_transaction_prohibits_use_of_commit_rollback_autocommit(conn):
-    """
-    Within a Transaction block, it is forbidden to touch commit, rollback,
-    or the autocommit setting on the connection, as this would interfere
-    with the transaction scope being managed by the Transaction block.
-    """
-    conn.autocommit = False
-    conn.commit()
-    conn.rollback()
-
-    with conn.transaction():
-        with pytest.raises(psycopg3.ProgrammingError):
-            conn.autocommit = False
-        with pytest.raises(psycopg3.ProgrammingError):
-            conn.commit()
-        with pytest.raises(psycopg3.ProgrammingError):
-            conn.rollback()
-
-    conn.autocommit = False
-    conn.commit()
-    conn.rollback()
-
-
-@pytest.mark.parametrize("autocommit", [False, True])
-def test_transaction_preserves_autocommit(conn, autocommit):
-    """
-    Connection.autocommit is unchanged both during and after Transaction block.
-    """
-    conn.autocommit = autocommit
-    with conn.transaction():
-        assert conn.autocommit is autocommit
-    assert conn.autocommit is autocommit
-
-
-def test_transaction_autocommit_off_but_no_transaction_started_successful_exit(
-    temp_table, conn, svcconn
-):
-    """
-    Scenario:
-    * Connection has autocommit off but no transaction has been initiated
-      before entering the Transaction context
-    * Code exits Transaction context successfully
-
-    Outcome:
-    * Changes made within Transaction context are committed
-    """
-    conn.autocommit = False
-    assert conn.pgconn.transaction_status == conn.TransactionStatus.IDLE
-    with conn.transaction():
-        insert_row(conn, "new")
-    assert conn.pgconn.transaction_status == conn.TransactionStatus.IDLE
-
-    # Changes committed
-    assert_rows(conn, {"new"})
-    assert_rows(svcconn, {"new"})
-
-
-def test_transaction_autocommit_off_but_no_transaction_started_exception_exit(
-    temp_table, conn, svcconn, exception=True
-):
-    """
-    Scenario:
-    * Connection has autocommit off but no transaction has been initiated
-      before entering the Transaction context
-    * Code exits Transaction context with an exception
-
-    Outcome:
-    * Changes made within Transaction context are discarded
-    """
-    conn.autocommit = False
-    assert conn.pgconn.transaction_status == conn.TransactionStatus.IDLE
-    with pytest.raises(ExpectedException):
-        with conn.transaction():
-            insert_row(conn, "new")
-            raise ExpectedException()
-    assert conn.pgconn.transaction_status == conn.TransactionStatus.IDLE
-
-    # Changes discarded
-    assert_rows(conn, set())
-    assert_rows(svcconn, set())
-
-
-def test_transaction_autocommit_off_and_transaction_in_progress_successful_exit(
-    temp_table, conn, svcconn
-):
-    """
-    Scenario:
-    * Connection has autocommit off but and a transaction is already in
-      progress before entering the Transaction context
-    * Code exits Transaction context successfully
-
-    Outcome:
-    * Changes made within Transaction context are left intact
-    * Outer transaction is left running, and no changes are visible to an
-      outside observer from another connection.
-    """
-    conn.autocommit = False
-    insert_row(conn, "prior")
-    assert conn.pgconn.transaction_status == conn.TransactionStatus.INTRANS
-    with conn.transaction():
-        insert_row(conn, "new")
-    assert conn.pgconn.transaction_status == conn.TransactionStatus.INTRANS
-    assert_rows(conn, {"prior", "new"})
-    # Nothing committed yet; changes not visible on another connection
-    assert_rows(svcconn, set())
-
-
-def test_transaction_autocommit_off_and_transaction_in_progress_exception_exit(
-    temp_table, conn, svcconn
-):
-    """
-    Scenario:
-    * Connection has autocommit off but and a transaction is already in
-      progress before entering the Transaction context
-    * Code exits Transaction context with an exception
-
-    Outcome:
-    * Changes made before the Transaction context are left intact
-    * Changes made within Transaction context are discarded
-    * Outer transaction is left running, and no changes are visible to an
-      outside observer from another connection.
-    """
-    conn.autocommit = False
-    insert_row(conn, "prior")
-    assert conn.pgconn.transaction_status == conn.TransactionStatus.INTRANS
-    with pytest.raises(ExpectedException):
-        with conn.transaction():
-            insert_row(conn, "new")
-            raise ExpectedException()
-    assert conn.pgconn.transaction_status == conn.TransactionStatus.INTRANS
-    assert_rows(conn, {"prior"})
-    # Nothing committed yet; changes not visible on another connection
-    assert_rows(svcconn, set())
-
-
-def test_transaction_nested_all_changes_persisted_on_successful_exit(
-    temp_table, conn, svcconn
-):
-    """Changes from nested transaction contexts are all persisted on exit."""
-    with conn.transaction():
-        insert_row(conn, "outer-before")
-        with conn.transaction():
-            insert_row(conn, "inner")
-        insert_row(conn, "outer-after")
-    assert conn.pgconn.transaction_status == conn.TransactionStatus.IDLE
-    assert_rows(conn, {"outer-before", "inner", "outer-after"})
-    assert_rows(svcconn, {"outer-before", "inner", "outer-after"})
-
-
-def test_transaction_nested_all_changes_discarded_on_outer_exception(
-    temp_table, conn, svcconn
-):
-    """
-    Changes from nested transaction contexts are discarded when an exception
-    raised in outer context escapes.
-    """
-    with pytest.raises(ExpectedException):
-        with conn.transaction():
-            insert_row(conn, "outer")
-            with conn.transaction():
-                insert_row(conn, "inner")
-            raise ExpectedException()
-    assert conn.pgconn.transaction_status == conn.TransactionStatus.IDLE
-    assert_rows(conn, set())
-    assert_rows(svcconn, set())
-
-
-def test_transaction_nested_all_changes_discarded_on_inner_exception(
-    temp_table, conn, svcconn
-):
-    """
-    Changes from nested transaction contexts are discarded when an exception
-    raised in inner context escapes the outer context.
-    """
-    with pytest.raises(ExpectedException):
-        with conn.transaction():
-            insert_row(conn, "outer")
-            with conn.transaction():
-                insert_row(conn, "inner")
-                raise ExpectedException()
-    assert conn.pgconn.transaction_status == conn.TransactionStatus.IDLE
-    assert_rows(conn, set())
-    assert_rows(svcconn, set())
-
-
-def test_transaction_nested_inner_scope_exception_handled_in_outer_scope(
-    temp_table, conn, svcconn
-):
-    """
-    An exception escaping the inner transaction context causes changes made
-    within that inner context to be discarded, but the error can then be
-    handled in the outer context, allowing changes made in the outer context
-    (both before, and after, the inner context) to be successfully committed.
-    """
-    with conn.transaction():
-        insert_row(conn, "outer-before")
-        with pytest.raises(ExpectedException):
-            with conn.transaction():
-                insert_row(conn, "inner")
-                raise ExpectedException()
-        insert_row(conn, "outer-after")
-    assert conn.pgconn.transaction_status == conn.TransactionStatus.IDLE
-    assert_rows(conn, {"outer-before", "outer-after"})
-    assert_rows(svcconn, {"outer-before", "outer-after"})
-
-
-def test_transaction_nested_three_levels_successful_exit(
-    temp_table, conn, svcconn
-):
-    with conn.transaction():
-        insert_row(conn, "one")
-        with conn.transaction():
-            insert_row(conn, "two")
-            with conn.transaction():
-                insert_row(conn, "three")
-    assert conn.pgconn.transaction_status == conn.TransactionStatus.IDLE
-    assert_rows(conn, {"one", "two", "three"})
-    assert_rows(svcconn, {"one", "two", "three"})
-
-
-def insert_row(conn, value):
-    conn.cursor().execute("INSERT INTO temp_table VALUES (%s)", (value,))
-
-
-def assert_rows(conn, expected):
-    rows = conn.cursor().execute("SELECT * FROM temp_table").fetchall()
-    assert set(v for (v,) in rows) == expected
-
-
-def assert_not_in_transaction(conn):
-    assert conn.pgconn.transaction_status == conn.TransactionStatus.IDLE
-
-
-def assert_in_transaction(conn):
-    assert conn.pgconn.transaction_status == conn.TransactionStatus.INTRANS
-
-
-class ExpectedException(Exception):
-    pass
